flask/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
icecream==2.1.1
simpletransformers==0.60.9
pyyaml==5.4.1
mojimoji==0.0.11
fugashi==1.1.0 
ipadic==1.0.0 
pytorch-lightning==1.4.5
transformers==4.10.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0